refinement: Update MessageWatch for aggregated libraries
Changed MessageWatch so that the loaded libraries are shown
as the tooltiptext of the status line and the table area.

Now the status text just states whether or not the libraries
are loaded. This is necessary since with aggregation their
may be many providers, so the status line would be too large
and become cumbersome.