bug: Fix for Test Manager.
The class field copyImage is intentionally not disposed now. If dispose is
called on it, it is then disposed in the shared registry that it is pulled
from. Then the next time the HostSelectionTable gets instantiated within
the same OTE-IDE session, it will still be disposed and cause an
exception.

Change-Id: I80c9d91198dd2275f02d1c254bc7b7018d607d90