feature: Adding test navigation to top level Test Point
When a test point was added via a promptPassFail or logTestPoint call, there
was no option to navigate to the line in the test source.  This change adds
a better filter to the stacktrace and location items so that they still aren't
allowing duplicates but are still allowed on top level Test Point items.

Change-Id: Id06782055572687ff70b013443d29d704be5b6db